Conflicts:
Not interesting:
	extras/Projucer/Builds/MacOSX/Icon.icns
	extras/Projucer/JuceLibraryCode/BinaryData.cpp

Changed because of our flag removed below
	extras/Projucer/Source/Project/jucer_Project.cpp

Not-really-conflict, we removed a flag and so did JUCE around same place
	extras/Projucer/Source/ProjectSaving/jucer_ProjectExport_Xcode.h

Not-really-conflicts due to our EnhancedAudioSuite additions
	modules/juce_audio_plugin_client/AAX/juce_AAX_Wrapper.cpp

Celemony refactors not-really-conflicts with our additions, merged
	modules/juce_audio_plugin_client/ARA/juce_ARAModelObjects.h

Changed by us to add IGNORE_MOUSE_WITH_PRO_TOOLS_AUTOMATION_MODIFIERS
Code moved to modules/juce_audio_processors/utilities/juce_ParameterAttachments, reapplied there
	modules/juce_audio_processors/utilities/juce_AudioProcessorValueTreeState.cpp

Changed by us in commits 05dd4310b1eb8b41e532f6947bcecc2fa29ff274s, 9db01c43626fabf0eef5ba19e5398b142a554dd3
Removed in juce6 but fixed instead in commits 05dd4310b1eb8b41e532f6947bcecc2fa29ff274s, 28d8f0d9e09021e9b3adc5ab99618b68816dbb6e
	modules/juce_gui_basics/native/juce_linux_X11_Clipboard.cpp
	modules/juce_gui_basics/native/juce_linux_X11_Windowing.cpp

+++ examples/GUI/AnimationDemo.h
@@ -123,10 +123,10 @@ struct BallComponent  : public Component
     void paint (Graphics& g) override
     {
         g.setColour (colour);
-        g.fillEllipse (2.0f, 2.0f, getWidth() - 4.0f, getHeight() - 4.0f);
+        g.fillEllipse (2.0f, 2.0f, (float) getWidth() - 4.0f, (float) getHeight() - 4.0f);
 
         g.setColour (Colours::darkgrey);
-        g.drawEllipse (2.0f, 2.0f, getWidth() - 4.0f, getHeight() - 4.0f, 1.0f);
+        g.drawEllipse (2.0f, 2.0f, (float) getWidth() - 4.0f, (float) getHeight() - 4.0f, 1.0f);
     }
 
     Point<float> position, speed;
@@ -262,10 +262,10 @@ private:
             auto newIndex = (componentsToAnimate.indexOf (component) + 3 * cycleCount)
                              % componentsToAnimate.size();
 
-            auto angle = newIndex * MathConstants<float>::twoPi / componentsToAnimate.size();
+            auto angle = (float) newIndex * MathConstants<float>::twoPi / (float) componentsToAnimate.size();
 
-            auto radius = useWidth ? width  * 0.35f
-                                   : height * 0.35f;
+            auto radius = useWidth ? (float) width  * 0.35f
+                                   : (float) height * 0.35f;
 
             Rectangle<int> r (getWidth()  / 2 + (int) (radius * std::sin (angle)) - 50,
                               getHeight() / 2 + (int) (radius * std::cos (angle)) - 50,
